[pigeon] Fixes swift int64 support for Swift (#3455)
[pigeon] Fixes swift int64 support for Swift

+++ packages/pigeon/platform_tests/test_plugin/macos/Classes/TestPlugin.swift
@@ -44,7 +44,7 @@ public class TestPlugin: NSObject, FlutterPlugin, HostIntegrationCoreApi {
     throw FlutterError(code: "code", message: "message", details: "details")
   }
 
-  func echo(_ anInt: Int32) -> Int32 {
+  func echo(_ anInt: Int64) -> Int64 {
     return anInt
   }
 
@@ -84,12 +84,12 @@ public class TestPlugin: NSObject, FlutterPlugin, HostIntegrationCoreApi {
     return AllNullableTypesWrapper(values: AllNullableTypes(aNullableString: nullableString))
   }
 
-  func sendMultipleNullableTypes(aBool aNullableBool: Bool?, anInt aNullableInt: Int32?, aString aNullableString: String?) -> AllNullableTypes {
+  func sendMultipleNullableTypes(aBool aNullableBool: Bool?, anInt aNullableInt: Int64?, aString aNullableString: String?) -> AllNullableTypes {
     let someThings = AllNullableTypes(aNullableBool: aNullableBool, aNullableInt: aNullableInt, aNullableString: aNullableString)
     return someThings
   }
 
-  func echo(_ aNullableInt: Int32?) -> Int32? {
+  func echo(_ aNullableInt: Int64?) -> Int64? {
     return aNullableInt
   }
 
@@ -141,7 +141,7 @@ public class TestPlugin: NSObject, FlutterPlugin, HostIntegrationCoreApi {
     completion(.success(everything))
   }
 
-  func echoAsync(_ anInt: Int32, completion: @escaping (Result<Int32, Error>) -> Void) {
+  func echoAsync(_ anInt: Int64, completion: @escaping (Result<Int64, Error>) -> Void) {
     completion(.success(anInt))
   }
 
@@ -173,7 +173,7 @@ public class TestPlugin: NSObject, FlutterPlugin, HostIntegrationCoreApi {
     completion(.success(aMap))
   }
 
-  func echoAsyncNullable(_ anInt: Int32?, completion: @escaping (Result<Int32?, Error>) -> Void) {
+  func echoAsyncNullable(_ anInt: Int64?, completion: @escaping (Result<Int64?, Error>) -> Void) {
     completion(.success(anInt))
   }
 
@@ -229,7 +229,7 @@ public class TestPlugin: NSObject, FlutterPlugin, HostIntegrationCoreApi {
 
   func callFlutterSendMultipleNullableTypes(
     aBool aNullableBool: Bool?,
-    anInt aNullableInt: Int32?,
+    anInt aNullableInt: Int64?,
     aString aNullableString: String?,
     completion: @escaping (Result<AllNullableTypes, Error>) -> Void
   ) {
@@ -248,7 +248,7 @@ public class TestPlugin: NSObject, FlutterPlugin, HostIntegrationCoreApi {
     }
   }
 
-  func callFlutterEcho(_ anInt: Int32, completion: @escaping (Result<Int32, Error>) -> Void) {
+  func callFlutterEcho(_ anInt: Int64, completion: @escaping (Result<Int64, Error>) -> Void) {
     flutterAPI.echo(anInt) {
       completion(.success($0))
     }
@@ -290,7 +290,7 @@ public class TestPlugin: NSObject, FlutterPlugin, HostIntegrationCoreApi {
     }
   }
 
-  func callFlutterEchoNullable(_ anInt: Int32?, completion: @escaping (Result<Int32?, Error>) -> Void) {
+  func callFlutterEchoNullable(_ anInt: Int64?, completion: @escaping (Result<Int64?, Error>) -> Void) {
     flutterAPI.echoNullable(anInt) {
       completion(.success($0))
     }
